remove unnecessary boxing

+++ src/main/java/de/katzenpapst/amunra/mob/entity/EntityAlienBug.java
@@ -145,7 +145,7 @@ public class EntityAlienBug extends EntityMob implements IEntityNonOxygenBreathe
             b0 &= -2;
         }
 
-        this.dataWatcher.updateObject(16, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(16, b0);
     }
 
     @Override
